Distribution Data

A data frame of information on the implemented distributions.

dist_data

Format

An object of class tbl_df (inherits from tbl, data.frame) with 10 rows and 4 columns.

Details

dist

The distribution (chr).

npars

The number of parameters (int).

tails

Whether the distribution has both tails (flag).

stable

Whether the distribution is numerically stable (flag).

bcanz

Whether the distribution belongs to the set of distributions approved by BC, Canada, Australia and New Zealand for official guidelines (flag).

See also

Examples

dist
#> function (x, method = "euclidean", diag = FALSE, upper = FALSE, 
#>     p = 2) 
#> {
#>     if (!is.na(pmatch(method, "euclidian"))) 
#>         method <- "euclidean"
#>     METHODS <- c("euclidean", "maximum", "manhattan", "canberra", 
#>         "binary", "minkowski")
#>     method <- pmatch(method, METHODS)
#>     if (is.na(method)) 
#>         stop("invalid distance method")
#>     if (method == -1) 
#>         stop("ambiguous distance method")
#>     x <- as.matrix(x)
#>     N <- nrow(x)
#>     attrs <- if (method == 6L) 
#>         list(Size = N, Labels = dimnames(x)[[1L]], Diag = diag, 
#>             Upper = upper, method = METHODS[method], p = p, call = match.call(), 
#>             class = "dist")
#>     else list(Size = N, Labels = dimnames(x)[[1L]], Diag = diag, 
#>         Upper = upper, method = METHODS[method], call = match.call(), 
#>         class = "dist")
#>     .Call(C_Cdist, x, method, attrs, p)
#> }
#> <bytecode: 0x55b0b651ddb8>
#> <environment: namespace:stats>