Parameter Descriptions for ssdtools Functions

Parameter Descriptions for ssdtools Functions

Arguments

all

A flag specifying whether to also return transformed parameters.

at_boundary_ok

A flag specifying whether a model with one or more parameters at the boundary should be considered to have converged (default = FALSE).

x

The object.

object

The object.

control

A list of control parameters passed to stats::optim().

chk

A flag specifying whether to check the arguments.

data

A data frame.

pred

A data frame of the predictions.

xlab

A string of the x-axis label.

ylab

A string of the x-axis label.

xbreaks

The x-axis breaks as one of:

NULL for no breaks

waiver() for the default breaks

A numeric vector of positions

breaks

A character vector

bounds

A named non-negative numeric vector of the left and right bounds for uncensored missing (0 and Inf) data in terms of the orders of magnitude relative to the extremes for non-missing values.

conc

A numeric vector of concentrations.

digits

A whole number specifying the number of significant figures

percent

A numeric vector of percentages.

pvalue

A flag specifying whether to return p-values or the statistics (default) for the various tests.

parametric

A flag specifying whether to perform parametric as opposed to non-parametric bootstrapping.

min_pmix

A number between 0 and 0.5 specifying the minimum proportion in mixture models.

delta

A non-negative number specifying the maximum absolute Akaike Information-theoretic Criterion difference cutoff. Distributions with an absolute difference from the best model greater than the cutoff are excluded.

average

A flag specifying whether to model average the estimates.

ci

A flag specifying whether to estimate confidence intervals (by parametric bootstrapping).

nboot

A count of the number of bootstrap samples to use to estimate the se and confidence limits. A value of 10000 is recommended for official guidelines.

min_pboot

A number of the minimum proportion of bootstrap samples that must successfully fit in the sense of returning a likelihood.

level

A number between 0 and 1 of the confidence level.

...

Unused.

q

vector of quantiles.

p

vector of probabilities.

n

number of observations.

log

logical; if TRUE, probabilities p are given as log(p).

log.p

logical; if TRUE, probabilities p are given as log(p).

lower.tail

logical; if TRUE (default), probabilities are P[X <= x],otherwise, P[X > x].

location

location parameter.

llocation

location parameter on the log scale.

scale

scale parameter.

lscale

scale parameter on the log scale.

shape1

shape1 parameter.

shape2

shape2 parameter.

range_shape1

A numeric vector of length two of the lower and upper bounds for the shape1 parameter.

range_shape2

shape2 parameter.

locationlog1

locationlog1 parameter.

scalelog1

scalelog1 parameter.

locationlog2

locationlog2 parameter.

scalelog2

scalelog2 parameter.

pmix

Proportion mixture parameter.

meanlog

mean on log scale parameter.

meanlog1

mean on log scale parameter.

meanlog2

mean on log scale parameter.

locationlog

location on log scale parameter.

sdlog

standard deviation on log scale parameter.

sdlog1

standard deviation on log scale parameter.

sdlog2

standard deviation on log scale parameter.

scalelog

scale on log scale parameter.

lshape

shape parameter on the log scale.

lshape1

shape1 parameter on the log scale.

lshape2

shape2 parameter on the log scale.

xintercept

The x-value for the intersect

yintercept

The y-value for the intersect.

select

A character vector of the distributions to select.

rescale

A flag specifying whether to rescale concentration values by dividing by the largest finite value.

reweight

A flag specifying whether to reweight weights by dividing by the largest weight.

left

A string of the column in data with the concentrations.

right

A string of the column in data with the right concentration values.

label

A string of the column in data with the labels.

shape

A string of the column in data for the shape aesthetic.

color

A string of the column in data for the color aesthetic.

size

A number for the size of the labels.

ribbon

A flag indicating whether to plot the confidence interval as a grey ribbon as opposed to green solid lines.

shift_x

The value to multiply the label x values by.

hc

A count between 1 and 99 indicating the percent hazard concentration (or NULL).

weight

A string of the numeric column in data with positive weights less than or equal to 1,000 or NULL.

dists

A character vector of the distribution names.

computable

A flag specifying whether to only return fits with numerically computable standard errors.

silent

A flag indicating whether fits should fail silently.

na.rm

A flag specifying whether to silently remove missing values or remove them with a warning.

nrow

A positive whole number of the minimum number of non-missing rows.

nsim

A positive whole number of the number of simulations to generate.

linetype

A string of the column in pred to use for the linetype.

linecolor

A string of the column in pred to use for the line color.