Predict fitdists

Predict fitdists

# S3 method for fitdists
predict(
  object,
  percent = 1:99,
  ci = FALSE,
  level = 0.95,
  nboot = 1000,
  parallel = NULL,
  ncpus = 1,
  average = TRUE,
  ic = "aicc",
  ...
)

Arguments

object
The object.

percent
A numeric vector of percentages.

ci
A flag specifying whether to estimate confidence intervals (by parametric bootstrapping).

level
A number between 0 and 1 of the confidence level.

nboot
A count of the number of bootstrap samples to use to estimate the se and confidence limits.

parallel
A string specifying the type of parallel operation to be used ('no', 'snow' or 'multicore').

ncpus
A count of the number of parallel processes to use.

average
A flag specifying whether to model average the estimates.

ic
A string specifying which information-theoretic criterion ('aic', 'aicc' or 'bic') to use for model averaging .

...
Unused.

Examples

predict(boron_dists)

#> # A tibble: 99 x 6
#>    percent   est    se   lcl   ucl dist   
#>      <int> <dbl> <dbl> <dbl> <dbl> <chr>  
#>  1       1 0.379    NA    NA    NA average
#>  2       2 0.624    NA    NA    NA average
#>  3       3 0.856    NA    NA    NA average
#>  4       4 1.08     NA    NA    NA average
#>  5       5 1.31     NA    NA    NA average
#>  6       6 1.53     NA    NA    NA average
#>  7       7 1.75     NA    NA    NA average
#>  8       8 1.98     NA    NA    NA average
#>  9       9 2.21     NA    NA    NA average
#> 10      10 2.44     NA    NA    NA average
#> # … with 89 more rows