Goodness of Fit

Returns a tbl data frame with the following columns

dist

The distribution name (chr)

aic

Akaike's Information Criterion (dbl)

bic

Bayesian Information Criterion (dbl)

and if the data are non-censored

aicc

Akaike's Information Criterion corrected for sample size (dbl)

and if there are 8 or more samples

ad

Anderson-Darling statistic (dbl)

ks

Kolmogorov-Smirnov statistic (dbl)

cvm

Cramer-von Mises statistic (dbl)

In the case of an object of class fitdists the function also returns

delta

The Information Criterion differences (dbl)

weight

The Information Criterion weights (dbl)

where delta and weight are based on aic for censored data and aicc for non-censored data.

ssd_gof(x, ...)

# S3 method for fitdist
ssd_gof(x, ...)

# S3 method for fitdistcens
ssd_gof(x, ...)

# S3 method for fitdists
ssd_gof(x, ...)

Arguments

x
The object.

...
Unused.

Value

A tbl data frame of the gof statistics.

Methods (by class)

fitdist: Goodness of Fit

fitdistcens: Goodness of Fit

fitdists: Goodness of Fit

Examples

ssd_gof(boron_lnorm)

#> # A tibble: 1 x 7
#>   dist     ad    ks    cvm   aic  aicc   bic
#> * <chr> <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl>
#> 1 lnorm 0.507 0.107 0.0703  239.  240.  242.

ssd_gof(boron_dists)

#> # A tibble: 6 x 9
#>   dist        ad     ks    cvm   aic  aicc   bic delta weight
#> * <chr>    <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl>
#> 1 lnorm    0.507 0.107  0.0703  239.  240.  242. 1.42   0.133
#> 2 llog     0.487 0.0993 0.0595  241.  241.  244. 3.40   0.049
#> 3 gompertz 0.602 0.120  0.0822  238.  238.  240. 0      0.271
#> 4 lgumbel  0.829 0.158  0.134   244.  245.  247. 6.58   0.01 
#> 5 gamma    0.441 0.117  0.0555  238.  238.  240. 0.019  0.268
#> 6 weibull  0.435 0.117  0.0543  238.  238.  240. 0.014  0.269