BCANZ Hazard Concentrations

Gets hazard concentrations with confidence intervals that protect 1, 5, 10 and 20% of species using settings adopted by BC, Canada, Australia and New Zealand for official guidelines. This function can take several minutes to run with recommended 10,000 iterations.

Usage

ssd_hc_bcanz(x, nboot = 10000, min_pboot = 0.95)

Arguments

x

The object.

nboot

A count of the number of bootstrap samples to use to estimate the confidence limits. A value of 10,000 is recommended for official guidelines.

min_pboot

A number between 0 and 1 of the minimum proportion of bootstrap samples that must successfully fit (return a likelihood) to report the confidence intervals.

Value

A tibble of corresponding hazard concentrations.

See also

Examples

fits <- ssd_fit_bcanz(ssddata::ccme_boron)
ssd_hc_bcanz(fits, nboot = 100)
#> # A tibble: 4 × 11
#>   dist    proportion   est    se    lcl   ucl    wt method   nboot pboot samples
#>   <chr>        <dbl> <dbl> <dbl>  <dbl> <dbl> <dbl> <chr>    <dbl> <dbl> <I<lis>
#> 1 average       0.01 0.267 0.476 0.0596  1.73     1 paramet…   100     1 <dbl>  
#> 2 average       0.05 1.26  0.880 0.453   3.94     1 paramet…   100     1 <dbl>  
#> 3 average       0.1  2.38  1.22  1.11    5.89     1 paramet…   100     1 <dbl>  
#> 4 average       0.2  4.81  1.81  2.70    9.50     1 paramet…   100     1 <dbl>