Water Quality Guideline for British Columbia

Calculates the 5% Hazard Concentration for British Columbia after rescaling the data based on the log-logistic, log-normal and gamma distributions using the parametric bootstrap and AICc model averaging.

Usage

ssd_wqg_bc(data, left = "Conc")

Arguments

data

A data frame.

left

A string of the column in data with the concentrations.

Value

A tibble of the 5% hazard concentration with 95% confidence intervals.

Details

Returns a tibble the model averaged 5% hazard concentration with standard errors, 95% lower and upper confidence limits and the number of bootstrap samples as well as the proportion of bootstrap samples that successfully returned a likelihood (convergence of the bootstrap sample is not required).

See also

Examples

if (FALSE) {
ssd_wqg_bc(ssddata::ccme_boron)
}