ssdtools: Species Sensitivity Distributions

Species sensitivity distributions are cumulative probability distributions which are fitted to toxicity concentrations for different species as described by Posthuma et al.(2001) <isbn:9781566705783>. The ssdtools package uses Maximum Likelihood to fit distributions such as the gamma, log-logistic, log-normal and log-normal log-normal mixture. Multiple distributions can be averaged using Akaike Information Criteria. Confidence intervals on hazard concentrations and proportions are produced by parametric bootstrapping.

See also

Author

Maintainer: Joe Thorley joe@poissonconsulting.ca (ORCID)

Authors:

Rebecca Fisher R.Fisher@aims.gov.au

David Fox david.fox@environmetrics.net.au

Carl Schwarz

Other contributors:

Angeline Tillmanns [contributor]

Seb Dalgarno seb@poissonconsulting.ca (ORCID) [contributor]

Kathleen McTavish [contributor]

Heather Thompson [contributor]

Doug Spry [contributor]

Rick van Dam [contributor]

Graham Batley [contributor]

Yulia Cuthbertson [contributor]

Tony Bigwood [contributor]

Michael Antenucci [contributor]

Ali Azizishirazi [contributor]

Nadine Hussein nadine@poissonconsulting.ca (ORCID) [contributor]

Sarah Lyons sarah@poissonconsulting.ca (ORCID) [contributor]

Stephanie Hazlitt [contributor]

Hadley Wickham [contributor]

Sergio Ibarra Espinosa [contributor]

Andy Teucher [contributor]

Emilie Doussantousse [contributor]

Nan-Hung Hsieh [contributor]

Province of British Columbia [funder, copyright holder]

Environment and Climate Change Canada [funder, copyright holder]

Australian Government Department of Climate Change, Energy, the Environment and Water [funder, copyright holder]