Distribution Data

A data frame of information on the implemented distributions.

Usage

dist_data

Format

An object of class tbl_df (inherits from tbl, data.frame) with 10 rows and 4 columns.

Details

dist

The distribution (chr).

npars

The number of parameters (int).

tails

Whether the distribution has both tails (flag).

stable

Whether the distribution is numerically stable (flag).

bcanz

Whether the distribution belongs to the set of distributions approved by BC, Canada, Australia and New Zealand for official guidelines (flag).

See also

Examples

dist_data
#> # A tibble: 10 × 4
#>    dist          bcanz tails npars
#>    <chr>         <lgl> <lgl> <int>
#>  1 burrIII3      FALSE TRUE      3
#>  2 gamma         TRUE  TRUE      2
#>  3 gompertz      FALSE TRUE      2
#>  4 invpareto     FALSE FALSE     2
#>  5 lgumbel       TRUE  TRUE      2
#>  6 llogis        TRUE  TRUE      2
#>  7 llogis_llogis FALSE TRUE      5
#>  8 lnorm         TRUE  TRUE      2
#>  9 lnorm_lnorm   TRUE  TRUE      5
#> 10 weibull       TRUE  TRUE      2