Parameter Descriptions for ssdtools Functions

Parameter Descriptions for ssdtools Functions

Arguments

x
The object.

object
The object.

data
A data frame.

pred
A data frame of the predictions.

xlab
A string of the x-axis label.

ylab
A string of the x-axis label.

conc
A numeric vector of concentrations.

percent
A numeric vector of percentages.

ic
A string specifying which information-theoretic criterion ('aic', 'aicc' or 'bic') to use for model averaging .

average
A flag specifying whether to model average the estimates.

ci
A flag specifying whether to estimate confidence intervals (by parametric bootstrapping).

nboot
A count of the number of bootstrap samples to use to estimate the se and confidence limits.

level
A number between 0 and 1 of the confidence level.

parallel
A string specifying the type of parallel operation to be used ('no', 'snow' or 'multicore').

ncpus
A count of the number of parallel processes to use.

...
Unused.

q
vector of quantiles.

p
vector of probabilities.

n
number of observations.

log, log.p
logical; if TRUE, probabilities p are given as log(p).

lower.tail
logical; if TRUE (default), probabilities are P[X <= x],otherwise, P[X > x].

location
location parameter.

llocation
location parameter on the log scale.

scale
scale parameter.

lscale
scale parameter on the log scale.

lshape
shape parameter on the log scale.

lshape1
shape1 parameter on the log scale.

lshape2
shape2 parameter on the log scale.

xintercept
The x-value for the intersect

yintercept
The y-value for the intersect.

select
A character vector of the distributions to select.

left
A string of the column in data with the concentrations.

right
A string of the column in data with the right concentration values.

label
A string of the column in data with the labels.

shape
A string of the column in data for the shape aesthetic.

color
A string of the column in data for the color aesthetic.

size
A number for the size of the labels.

ribbon
A flag indicating whether to plot the confidence interval as a grey ribbon as opposed to green solid lines.

shift_x
The value to multiply the label x values by.

hc
A count between 1 and 99 indicating the percent hazard concentration (or NULL).

weight
A string of the column in data with the weightings (or NULL)

dists
A character vector of the distributions to fit.

computable
A flag specifying whether to only return fits with numerically computable standard errors.

silent
A flag indicating whether fits should fail silently.

na.rm
A flag specifying whether to silently remove missing values or remove them with a warning.