Fit Distributions

Fits one or more distributions to species sensitivity data.

ssd_fit_dists(data, left = "Conc", right = left, weight = NULL,
  dists = c("lnorm", "llog", "gompertz", "lgumbel", "gamma", "weibull"),
  silent = FALSE)

Arguments

data
A data frame.

left
A string of the column in data with the left concentration values.

right
A string of the column in data with the right concentration values.

weight
A string of the column in data with the weightings (or NULL)

dists
A character vector of the distributions to fit.

silent
A flag indicating whether fits should fail silently.

Value

An object of class fitdists (a list of fitdist objects).

Details

By default the 'lnorm', 'llog', 'gompertz', 'lgumbel', 'gamma' and 'weibull' distributions are fitted to the data. The ssd_fit_dist and ssd_fit_dist functions have also been tested with the 'pareto' distribution.

If weight specifies a column in the data frame with positive integers, weighted estimation occurs. However, currently only the resultant parameter estimates are available (via coef).

If the `right` argument is different to the `left` argument then the data are considered to be censored. It may be possible to use artificial censoring to improve the estimates in the extreme tails (Liu et al 2018).

References

Liu, Y., Salibián-Barrera, M., Zamar, R.H., and Zidek, J.V. 2018. Using artificial censoring to improve extreme tail quantile estimates. Journal of the Royal Statistical Society: Series C (Applied Statistics).

See also

Examples

ssd_fit_dists(boron_data)

#> Fitting of the distribution ' lnorm ' by maximum likelihood 
#> Parameters:
#>         estimate Std. Error
#> meanlog 2.561645  0.2346291
#> sdlog   1.241540  0.1659073
#> Fitting of the distribution ' llog ' by maximum likelihood 
#> Parameters:
#>        estimate Std. Error
#> shape 0.7404232  0.1143730
#> scale 2.6261311  0.2482597
#> Fitting of the distribution ' gompertz ' by maximum likelihood 
#> Parameters:
#>          estimate  Std. Error
#> shape 0.039409997 0.011958000
#> scale 0.002601631 0.009957107
#> Fitting of the distribution ' lgumbel ' by maximum likelihood 
#> Parameters:
#>          estimate Std. Error
#> location 1.922498  0.2473440
#> scale    1.232358  0.1730647
#> Fitting of the distribution ' gamma ' by maximum likelihood 
#> Parameters:
#>        estimate Std. Error
#> scale 25.108064  7.6313816
#> shape  0.950757  0.2226832
#> Fitting of the distribution ' weibull ' by maximum likelihood 
#> Parameters:
#>         estimate Std. Error
#> shape  0.9662825  0.1454446
#> scale 23.5097478  4.8528313

data(fluazinam, package = "fitdistrplus")
ssd_fit_dists(fluazinam, left = "left", right = "right")

#> Warning: gompertz failed to fit: Error in if (one.more) { : missing value where TRUE/FALSE needed

#> Fitting of the distribution ' lnorm ' on censored data by maximum likelihood 
#> Parameters:
#>         estimate
#> meanlog 4.976920
#> sdlog   2.687785
#> Fitting of the distribution ' llog ' on censored data by maximum likelihood 
#> Parameters:
#>       estimate
#> shape 1.591448
#> scale 4.954329
#> Fitting of the distribution ' lgumbel ' on censored data by maximum likelihood 
#> Parameters:
#>          estimate
#> location 3.758518
#> scale    2.633732
#> Fitting of the distribution ' gamma ' on censored data by maximum likelihood 
#> Parameters:
#>           estimate
#> scale 2445.7299203
#> shape    0.3464374
#> Fitting of the distribution ' weibull ' on censored data by maximum likelihood 
#> Parameters:
#>          estimate
#> shape   0.4411236
#> scale 434.4368078