Fit Distributions

Fits one or more distributions to species sensitivity data.

ssd_fit_dists(
  data,
  left = "Conc",
  right = left,
  weight = NULL,
  dists = c("burrIII2", "gamma", "lnorm"),
  computable = TRUE,
  silent = FALSE
)

Arguments

data
A data frame.

left
A string of the column in data with the concentrations.

right
A string of the column in data with the right concentration values.

weight
A string of the column in data with the weightings (or NULL)

dists
A character vector of the distributions to fit.

computable
A flag specifying whether to only return fits with numerically computable standard errors.

silent
A flag indicating whether fits should fail silently.

Value

An object of class fitdists (a list of fitdist objects).

Details

By default the 'burrIII2', 'gamma' and 'lnorm' distributions are fitted to the data. The ssd_fit_dists function has also been tested with the 'burrIII3', 'gompertz', 'lgumbel', 'llogis', 'pareto' and 'weibull' distributions.

If weight specifies a column in the data frame with positive integers, weighted estimation occurs. However, currently only the resultant parameter estimates are available (via coef).

If the `right` argument is different to the `left` argument then the data are considered to be censored. It may be possible to use artificial censoring to improve the estimates in the extreme tails (Liu et al 2018).

References

Liu, Y., Salibián-Barrera, M., Zamar, R.H., and Zidek, J.V. 2018. Using artificial censoring to improve extreme tail quantile estimates. Journal of the Royal Statistical Society: Series C (Applied Statistics).

Examples

ssd_fit_dists(boron_data)

#> Warning: The `dists` argument of `ssd_fit_dists()` is deprecated as of lifecycle 0.1.0.
#> More specifically the default value of c('gamma', 'gompertz', 'lgumbel', 'llogis', 'lnorm', 'weibull') has been replaced by c('burrIII2', 'gamma', 'lnorm').
#> This warning is displayed once per session.
#> Call `lifecycle::last_warnings()` to see where this warning was generated.

#> Fitting of the distribution ' burrIII2 ' by maximum likelihood 
#> Parameters:
#>          estimate Std. Error
#> lshape  0.2076248  0.1134175
#> lscale -2.3663820  0.2798350
#> Fitting of the distribution ' gamma ' by maximum likelihood 
#> Parameters:
#>         estimate Std. Error
#> scale 25.1263769  7.6375478
#> shape  0.9500513  0.2224988
#> Fitting of the distribution ' lnorm ' by maximum likelihood 
#> Parameters:
#>         estimate Std. Error
#> meanlog 2.561645  0.2346291
#> sdlog   1.241540  0.1659073

data(fluazinam, package = "fitdistrplus")
ssd_fit_dists(fluazinam, left = "left", right = "right")

#> Fitting of the distribution ' burrIII2 ' on censored data by maximum likelihood 
#> Parameters:
#>          estimate
#> lshape -0.3068955
#> lscale -5.6156439
#> Fitting of the distribution ' gamma ' on censored data by maximum likelihood 
#> Parameters:
#>           estimate
#> scale 2446.6063266
#> shape    0.3465141
#> Fitting of the distribution ' lnorm ' on censored data by maximum likelihood 
#> Parameters:
#>         estimate
#> meanlog 4.976920
#> sdlog   2.687785