Fit Distributions

Fits one or more distributions to species sensitivity data.

ssd_fit_dists(
  data,
  left = "Conc",
  right = left,
  weight = NULL,
  dists = c("llogis", "gamma", "lnorm"),
  computable = FALSE,
  silent = FALSE
)

Arguments

data
A data frame.

left
A string of the column in data with the concentrations.

right
A string of the column in data with the right concentration values.

weight
A string of the column in data with the weightings (or NULL)

dists
A character vector of the distribution names.

computable
A flag specifying whether to only return fits with numerically computable standard errors.

silent
A flag indicating whether fits should fail silently.

Value

An object of class fitdists (a list of fitdist objects).

Details

By default the 'llogis', 'gamma' and 'lnorm' distributions are fitted to the data. The ssd_fit_dists function has also been tested with the 'burrIII3', 'gompertz', 'lgumbel' and 'weibull' distributions.

If weight specifies a column in the data frame with positive integers, weighted estimation occurs. However, currently only the resultant parameter estimates are available (via coef).

If the `right` argument is different to the `left` argument then the data are considered to be censored.

The fits are performed using fitdist (and fitdistcens in the case of censored data). The method used is "mle" (maximum likelihood estimation) which means that numerical optimization is carried out in mledist using optim unless finite bounds are supplied in the (lower and upper) in which it is carried out using constrOptim. In both cases the "Nelder-Mead" method is used.

Examples

ssd_fit_dists(boron_data)

#> Fitting of the distribution ' llogis ' by maximum likelihood 
#> Parameters:
#>         estimate Std. Error
#> lscale 2.6261249  0.2482332
#> lshape 0.7403092  0.1143365
#> Fitting of the distribution ' gamma ' by maximum likelihood 
#> Parameters:
#>         estimate Std. Error
#> scale 25.1263769  7.6375478
#> shape  0.9500513  0.2224988
#> Fitting of the distribution ' lnorm ' by maximum likelihood 
#> Parameters:
#>         estimate Std. Error
#> meanlog 2.561645  0.2346291
#> sdlog   1.241540  0.1659073

data(fluazinam, package = "fitdistrplus")
ssd_fit_dists(fluazinam, left = "left", right = "right")

#> Fitting of the distribution ' llogis ' on censored data by maximum likelihood 
#> Parameters:
#>        estimate
#> lscale 4.954683
#> lshape 1.591178
#> Fitting of the distribution ' gamma ' on censored data by maximum likelihood 
#> Parameters:
#>           estimate
#> scale 2446.6063266
#> shape    0.3465141
#> Fitting of the distribution ' lnorm ' on censored data by maximum likelihood 
#> Parameters:
#>         estimate
#> meanlog 4.976920
#> sdlog   2.687785