Hazard Concentrations for Species Sensitivity Distributions

Calculates concentration(s) with bootstrap confidence intervals that protect specified proportion(s) of species for individual or model-averaged distributions using parametric or non-parametric bootstrapping.

Usage

ssd_hc(x, ...)

# S3 method for class 'list'
ssd_hc(x, percent, proportion = 0.05, ...)

# S3 method for class 'fitdists'
ssd_hc(
  x,
  percent,
  proportion = 0.05,
  average = TRUE,
  ci = FALSE,
  level = 0.95,
  nboot = 1000,
  min_pboot = 0.95,
  multi_est = TRUE,
  ci_method = "weighted_samples",
  parametric = TRUE,
  delta = 9.21,
  samples = FALSE,
  save_to = NULL,
  control = NULL,
  ...
)

# S3 method for class 'fitburrlioz'
ssd_hc(
  x,
  percent,
  proportion = 0.05,
  ci = FALSE,
  level = 0.95,
  nboot = 1000,
  min_pboot = 0.95,
  parametric = FALSE,
  samples = FALSE,
  save_to = NULL,
  ...
)

Arguments

x

The object.

...

Unused.

percent

A numeric vector of percent values to estimate hazard concentrations for. Soft-deprecated for proportion = 0.05.

proportion

A numeric vector of proportion values to estimate hazard concentrations for.

average

A flag specifying whether to provide model averaged values as opposed to a value for each distribution.

ci

A flag specifying whether to estimate confidence intervals (by bootstrapping).

level

A number between 0 and 1 of the confidence level of the interval.

nboot

A count of the number of bootstrap samples to use to estimate the confidence limits. A value of 10,000 is recommended for official guidelines.

min_pboot

A number between 0 and 1 of the minimum proportion of bootstrap samples that must successfully fit (return a likelihood) to report the confidence intervals.

multi_est

A flag specifying whether to treat the distributions as constituting a single distribution (as opposed to taking the mean) when calculating model averaged estimates.

ci_method

A string specifying which method to use for estimating the bootstrap values. Possible values are "multi_free" and "multi_fixed" which treat the distributions as constituting a single distribution but differ in whether the model weights are fixed and "weighted_samples" and "weighted_arithmetic" take bootstrap samples from each distribution proportional to its weight versus calculating the weighted arithmetic means of the lower and upper confidence limits.

parametric

A flag specifying whether to perform parametric bootstrapping as opposed to non-parametrically resampling the original data with replacement.

delta

A non-negative number specifying the maximum absolute AIC difference cutoff. Distributions with an absolute AIC difference greater than delta are excluded from the calculations.

samples

A flag specfying whether to include a numeric vector of the bootstrap samples as a list column in the output.

save_to

NULL or a string specifying a directory to save where the bootstrap datasets and parameter estimates (when successfully converged) to.

control

A list of control parameters passed to stats::optim().

Value

A tibble of corresponding hazard concentrations.

Details

Model-averaged estimates and/or confidence intervals (including standard error) can be calculated by treating the distributions as constituting a single mixture distribution versus 'taking the mean'. When calculating the model averaged estimates treating the distributions as constituting a single mixture distribution ensures that ssd_hc() is the inverse of ssd_hp().

If treating the distributions as constituting a single mixture distribution when calculating model average confidence intervals then weighted specifies whether to use the original model weights versus re-estimating for each bootstrap sample unless 'taking the mean' in which case weighted specifies whether to take bootstrap samples from each distribution proportional to its weight (so that they sum to nboot) versus calculating the weighted arithmetic means of the lower and upper confidence limits based on nboot samples for each distribution.

Distributions with an absolute AIC difference greater than a delta of by default 7 have considerably less support (weight < 0.01) and are excluded prior to calculation of the hazard concentrations to reduce the run time.

Methods (by class)

ssd_hc(list): Hazard Concentrations for Distributional Estimates

ssd_hc(fitdists): Hazard Concentrations for fitdists Object

ssd_hc(fitburrlioz): Hazard Concentrations for fitburrlioz Object

References

Burnham, K.P., and Anderson, D.R. 2002. Model Selection and Multimodel Inference. Springer New York, New York, NY. doi:10.1007/b97636.

See also

Examples

ssd_hc(ssd_match_moments())
#> # A tibble: 6 × 9
#>   dist        proportion   est    se   lcl   ucl    wt nboot pboot
#>   <chr>            <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <int> <dbl>
#> 1 gamma             0.05 0.481    NA    NA    NA     1     0    NA
#> 2 lgumbel           0.05 0.739    NA    NA    NA     1     0    NA
#> 3 llogis            0.05 0.562    NA    NA    NA     1     0    NA
#> 4 lnorm             0.05 0.558    NA    NA    NA     1     0    NA
#> 5 lnorm_lnorm       0.05 0.469    NA    NA    NA     1     0    NA
#> 6 weibull           0.05 0.501    NA    NA    NA     1     0    NA

fits <- ssd_fit_dists(ssddata::ccme_boron)
ssd_hc(fits)
#> # A tibble: 1 × 11
#>   dist    proportion   est    se   lcl   ucl    wt method    nboot pboot samples
#>   <chr>        <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <chr>     <int> <dbl> <I<lis>
#> 1 average       0.05  1.26    NA    NA    NA     1 parametr…     0   NaN <dbl>  

fit <- ssd_fit_burrlioz(ssddata::ccme_boron)
ssd_hc(fit)
#> # A tibble: 1 × 11
#>   dist      proportion   est    se   lcl   ucl    wt method  nboot pboot samples
#>   <chr>          <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <chr>   <int> <dbl> <I<lis>
#> 1 invpareto       0.05 0.387    NA    NA    NA     1 non-pa…     0    NA <dbl>